debt - move integrity service to workbench
