+++ src/tsconfig.strictNullChecks.json
@@ -283,8 +283,6 @@
 		"./vs/platform/instantiation/test/common/graph.test.ts",
 		"./vs/platform/instantiation/test/common/instantiationService.test.ts",
 		"./vs/platform/instantiation/test/common/instantiationServiceMock.ts",
-		"./vs/platform/integrity/common/integrity.ts",
-		"./vs/platform/integrity/node/integrityServiceImpl.ts",
 		"./vs/platform/issue/common/issue.ts",
 		"./vs/platform/issue/electron-main/issueService.ts",
 		"./vs/platform/issue/node/issueIpc.ts",
@@ -706,6 +704,8 @@
 		"./vs/workbench/services/hash/node/hashService.ts",
 		"./vs/workbench/services/history/common/history.ts",
 		"./vs/workbench/services/history/electron-browser/history.ts",
+		"./vs/workbench/services/integrity/common/integrity.ts",
+		"./vs/workbench/services/integrity/node/integrityServiceImpl.ts",
 		"./vs/workbench/services/keybinding/common/keybindingIO.ts",
 		"./vs/workbench/services/keybinding/common/keyboardMapper.ts",
 		"./vs/workbench/services/keybinding/common/macLinuxFallbackKeyboardMapper.ts",
